% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyweight.R
\docType{data}
\name{bodyweight}
\alias{bodyweight}
\alias{bi}
\alias{cr}
\alias{ha}
\alias{kn}
\alias{lu}
\alias{mo}
\title{Bodyweight Exercise Data}
\format{
A tibble with 8 variables:
\describe{
\item{id}{ID of the athlete.}
\item{ex}{exercise type}
\item{set}{The set number of the exercise.}
\item{loc}{Location of body extremity.}
\item{dim}{Direction of rotation / acceleration}
\item{rep}{The count of the repetition.}
\item{acc}{Acceleration in \eqn{m/s^2}}
\item{rot}{Rotation in \eqn{rad/s}}
\item{rating}{Quality of the exercise (range of 1 (best) to 5)}
}
}
\source{
\url{https://github.com/andrebert/body-weight-exercises}
}
\usage{
bodyweight
}
\description{
Sensor-based measurements and quality ratings of athletes performing different
exercises, recorded across multiple sets and repetitions at different locations
on their bodies.

Argument values for acceleration and rotation are in ms since start of the repetition.

The data sets / exercises are:
\itemize{
\item \code{bi}: "Bicycle Crunch" (one rep is left side - right side),
\item \code{cr}: "Crunch",
\item \code{ha}: "Jumping Jack",
\item \code{kn}: "Squat",
\item \code{lu}: "Lunge",
\item \code{mo}: "Mountain Climber"
}
}
\examples{
tibble::glimpse(bi)
}
\references{
Perego, Paolo, Rahmani, M. A, TaheriNejad, Nima (eds.) (2018).
\emph{An Open, Labeled Dataset for Analysis and Assessment of Human Motion}.
ISBN 978-3-319-98551-0.
}
\keyword{datasets}
