% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chf.R
\docType{data}
\name{chf}
\alias{chf}
\title{Congestive heart failure accelerometry data}
\format{
A tibble with 13,300 rows and 9 variables:
\describe{
\item{id}{Unique identifier for each participant in the study.}
\item{gender}{Participant's gender, recorded as male or female.}
\item{age}{Age of the participant, in years.}
\item{bmi}{Body Mass Index (BMI) of the participant.}
\item{event_week}{The week number relative to a critical event.}
\item{event_type}{Type of event recorded, such as hospitalization.}
\item{week}{Week number during the study, indicating the time frame of data collection.}
\item{day}{Day of the week (e.g. Monday, Tuesday) when the activity data was recorded.}
\item{activity}{Activity level recorded by accelerometer (resolution is 1 minute, so 24 h/day * 60 min/h = 1440 min/day).}
}
}
\source{
\url{https://www.dropbox.com/scl/fi/vnuldgqnkfa532hzoazjj/chfData.zip}
}
\usage{
chf
}
\description{
Activity data from a study of congestive heart failure.
}
\examples{
chf
}
\references{
Huang, Lei, Bai, Jiawei, Ivanescu, Andrada, Harris, Tamara, Maurer, Mathew, Green, Philip, Zipunnikov, Vadim (2019).
\dQuote{Multilevel matrix-variate analysis and its application to accelerometry-measured physical activity in clinical populations.}
\emph{Journal of the American Statistical Association}.
}
\keyword{datasets}
