% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfundata-package.R
\docType{package}
\name{tidyfundata-package}
\alias{tidyfundata}
\alias{tidyfundata-package}
\title{tidyfundata: A Collection of Datasets Containing Functional Data}
\description{
A Collection of Datasets containing functional data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidyfun.github.io/tidyfundata/}
  \item \url{https://github.com/tidyfun/tidyfundata}
  \item Report bugs at \url{https://github.com/tidyfun/tidyfundata/issues}
}

}
\author{
\strong{Maintainer}: Fabian Scheipl \email{fabian.scheipl@googlemail.com} (\href{https://orcid.org/0000-0001-8172-3603}{ORCID})

Authors:
\itemize{
  \item Maximilian Mücke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})
}

}
\keyword{internal}
