% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyweight.R
\docType{data}
\name{bodyweight}
\alias{bodyweight}
\title{Bicycle Crunch data}
\format{
A tibble with 20,760 rows and 8 variables:
\describe{
\item{id}{ID of the athlete.}
\item{set}{The set number of the exercise.}
\item{loc}{Location of body extremity.}
\item{sens}{Indicator for acceleration (acc) or rotation (rot).}
\item{dim}{Dimension of rotation.}
\item{rep}{The number of repetitions.}
\item{rating}{Quality of their exercise sets within a range of 1 to 5.}
\item{activity}{The bicycle crunch activity.}
}
}
\source{
\url{https://github.com/andrebert/body-weight-exercises}
}
\usage{
bodyweight
}
\description{
Sensor-based measurements and quality ratings of athletes performing the
Bicycle Crunch exercise, recorded across multiple sets and repetitions.
}
\examples{
bodyweight
}
\references{
Perego, Paolo, Rahmani, M. A, TaheriNejad, Nima (eds.) (2018).
\emph{An Open, Labeled Dataset for Analysis and Assessment of Human Motion}.
ISBN 978-3-319-98551-0.
}
\keyword{datasets}
