% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dti}
\alias{dti}
\title{Diffusion tensor imaging data}
\format{
A tibble with 382 rows and 6 variables:
\describe{
\item{id}{}
\item{visit}{}
\item{sex}{}
\item{case}{}
\item{cca}{}
\item{rcst}{}
}
}
\usage{
dti
}
\description{
Fractional anisotropy (FA) tract profiles for the corpus
callosum (cca) and the right corticospinal tract (rcst). Accompanying
the tract profiles are the subject ID numbers, visit number, total
number of scans, multiple sclerosis case status and Paced Auditory Serial
Addition Test (pasat) score.

Data are also include in the \code{refund} package in another format.
}
\details{
If you use this data as an example in written work, please
include the following acknowledgment: “The MRI/DTI data were collected
at Johns Hopkins University and the Kennedy-Krieger Institute"
}
\examples{
dti
}
\keyword{datasets}
